:truck: [#4606] Move catalogue select to shared form library
Since Objects API and ZGW APIs registration options both make use of
this functionality, we can make it agnostic and move it into a
specialized form components directory for ZGW APIs interaction.
