+++ src/openforms/js/components/admin/form_design/registrations/objectsapi/fields/index.js
@@ -6,4 +6,3 @@ export {DocumentTypesFieldet} from './DocumentTypes';
 export {default as UpdateExistingObject} from './UpdateExistingObject';
 export {default as UploadSubmissionCsv} from './UploadSubmissionCSV';
 export {default as OrganisationRSIN} from './OrganisationRSIN';
-export {default as CatalogueSelect} from './CatalogueSelect';
