+++ src/openforms/js/components/admin/forms/zgw/CatalogueSelect.js
@@ -1,10 +1,19 @@
 import {useFormikContext} from 'formik';
 import PropTypes from 'prop-types';
-import {FormattedMessage} from 'react-intl';
 
 import Field from 'components/admin/forms/Field';
 import FormRow from 'components/admin/forms/FormRow';
 import ReactSelect from 'components/admin/forms/ReactSelect';
+
+/**
+ * @typedef {Object} Catalogue
+ * @property {string} rsin - The RSIN identifier of the organisation owning the catalogue.
+ * @property {string} label - A label to identify the catalogue in a dropdown.
+ *
+ * @typedef {Object} CatalogueGroup
+ * @property {string} label
+ * @property {Catalogue[]} options
+ */
 
 export const extractValue = (optionGroups, currentValue) => {
   const allOptions = optionGroups.reduce((acc, group) => acc.concat(group.options), []);
@@ -15,9 +24,32 @@ export const extractValue = (optionGroups, currentValue) => {
   );
 };
 
-const CatalogueSelect = ({loading, optionGroups}) => {
+/**
+ * Group options by organisation RSIN and sort them by label.
+ * @param  {Catalogue[]} catalogues List of catalogues, typically retrieved from an API endpoint.
+ * @return {CatalogueGroup[]} An array of catalogue groups. Each group contains the catalogues of each RSIN.
+ */
+export const groupAndSortOptions = catalogues => {
+  const _optionsByRSIN = {};
+  for (const catalogue of catalogues) {
+    const {rsin} = catalogue;
+    if (!_optionsByRSIN[rsin]) _optionsByRSIN[rsin] = [];
+    _optionsByRSIN[rsin].push(catalogue);
+  }
+
+  const groups = Object.entries(_optionsByRSIN)
+    .map(([rsin, options]) => ({
+      label: rsin,
+      options: options.sort((a, b) => a.label.localeCompare(b.label)),
+    }))
+    .sort((a, b) => a.label.localeCompare(b.label));
+
+  return groups;
+};
+
+const CatalogueSelect = ({label, isDisabled = false, loading, optionGroups}) => {
   const {
-    values: {objectsApiGroup = null, catalogue = {}},
+    values: {catalogue = {}},
     getFieldHelpers,
   } = useFormikContext();
 
@@ -25,21 +57,12 @@ const CatalogueSelect = ({loading, optionGroups}) => {
   const value = extractValue(optionGroups, catalogue);
   return (
     <FormRow>
-      <Field
-        name="catalogue"
-        label={
-          <FormattedMessage
-            description="Objects API registration options 'catalogue' label"
-            defaultMessage="Catalogue"
-          />
-        }
-        noManageChildProps
-      >
+      <Field name="catalogue" label={label} noManageChildProps>
         <ReactSelect
           name="catalogue"
           options={optionGroups}
           isLoading={loading}
-          isDisabled={!objectsApiGroup}
+          isDisabled={isDisabled}
           // override from the default because we're using an object as value
           value={value}
           onChange={selectedOption => {
@@ -56,6 +79,8 @@ const CatalogueSelect = ({loading, optionGroups}) => {
 };
 
 CatalogueSelect.propTypes = {
+  label: PropTypes.node.isRequired,
+  isDisabled: PropTypes.bool,
   loading: PropTypes.bool.isRequired,
   optionGroups: PropTypes.arrayOf(
     PropTypes.shape({
