+++ src/openforms/js/components/admin/forms/zgw/index.js
@@ -1,0 +1,8 @@
+/**
+ * Form fields related to the ZGW APIs standards/interaction.
+ */
+export {
+  default as CatalogueSelect,
+  extractValue as getCatalogueOption,
+  groupAndSortOptions as groupAndSortCatalogueOptions,
+} from './CatalogueSelect';
